Comment Creation & Load
Comments can be created on a project and comments load on a project. This currently only adds comments to one project and loads comments from every project. Once a projectId can be defined when a project is actually selected, then its a simple switch to change to a per-project basis.